Add info how to do dependent attributes
I had a factory for which I wanted some attributes to be derived from
the values of other attributes but I didn't know how to do it. I didn't
find the solution in the guide or API docs, so after digging into the
code I decided to add a little "how-to" to the guide.